refactor: move logic for calculating neibhbors out of the Grid class
Prefer composition over inheritance. And when I think about how
this project is likely to evolve, it dawns on me that it will be much
easier to switch between wrapped and bounded grid styles in the middle
of a step if I only have to switch NeighborSelector implementations, not
copy one Grid subclass's state into another.